requirements.txt
amqp==5.1.1
aniso8601==9.0.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==5.0.1
blinker==1.5
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==1.2.0
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.1
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
flask-logging-decorator==0.0.5
Flask-Navigation==0.2.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
fonttools==4.38.0
gunicorn==20.1.0
greenlet==2.0.1
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.7.2
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.4
kombu==5.2.4
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.3
mistune==0.8.4
mysql-connector-python==8.0.32
mysqlclient==2.1.1
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.4
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.3
psutil==5.9.1
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.6
pywin32==304
pywinpty==2.0.6
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
redis==4.5.3
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.1
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0.post1
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
sqlparse==0.4.3
stack-data==0.3.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tornado==6.2
traitlets==5.3.0
tzdata==2022.7
vine==5.0.0
virtualenv==20.16.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==3.6.1
